Create testeruino.ino
This file is Arduino format, and must be compiled with an IDE for Arduino. The code is a c++ spinoff. The Hardware it runs on is a PCB I made, and is used for in depth analysis of sensor, with the possibilities of fast combining resistors and various wiring configurations.